+++ api/addons/router.py
@@ -2,7 +2,7 @@ from typing import Any
 
 from fastapi import APIRouter
 
-from ayon_server.api import ResponseFactory
+from ayon_server.api.responses import ResponseFactory
 
 route_meta: dict[str, Any] = {
     "tags": ["Addon settings"],
