--- ayon_server/api/__init__.py
@@ -1,35 +1,0 @@
-__all__ = [
-    "JSONResponse",
-    "ResponseFactory",
-    "dep_access_token",
-    "dep_attribute_name",
-    "dep_current_user",
-    "dep_current_user_optional",
-    "dep_project_name",
-    "dep_new_project_name",
-    "dep_folder_id",
-    "dep_product_id",
-    "dep_version_id",
-    "dep_representation_id",
-    "dep_task_id",
-    "dep_thumbnail_id",
-    "dep_workfile_id",
-    "app",
-]
-
-from ayon_server.api.dependencies import (
-    dep_access_token,
-    dep_attribute_name,
-    dep_current_user,
-    dep_current_user_optional,
-    dep_folder_id,
-    dep_new_project_name,
-    dep_product_id,
-    dep_project_name,
-    dep_representation_id,
-    dep_task_id,
-    dep_thumbnail_id,
-    dep_version_id,
-    dep_workfile_id,
-)
-from ayon_server.api.responses import JSONResponse, ResponseFactory
